+++ cartoframes/context.py
@@ -59,10 +59,10 @@ DEFAULT_SQL_ARGS = dict(client='cartoframes_{}'.format(__version__),
 
 
 class CartoContext(object):
-    """CartoContext class for authentication with CARTO and high-level operations
-    such as reading tables from CARTO into dataframes, writing dataframes to
-    CARTO tables, creating custom maps from dataframes and CARTO tables, and
-    augmenting data using CARTO's `Data Observatory
+    """CartoContext class for authentication with CARTO and high-level
+    operations such as reading tables from CARTO into dataframes, writing
+    dataframes to CARTO tables, creating custom maps from dataframes and CARTO
+    tables, and augmenting data using CARTO's `Data Observatory
     <https://carto.com/data-observatory>`__. Future methods will interact with
     CARTO's services like `routing, geocoding, and isolines
     <https://carto.com/location-data-services/>`__, PostGIS backend for spatial
@@ -70,10 +70,26 @@ class CartoContext(object):
 
     Manages connections with CARTO for data and map operations. Modeled
     after `SparkContext
-    <https://jaceklaskowski.gitbooks.io/mastering-apache-spark/content/spark-sparkcontext.html>`__.
+    <http://spark.apache.org/docs/2.1.0/api/python/pyspark.html#pyspark.SparkContext>`__.
+
+    There are two ways of authenticating against a CARTO account:
+
+      1. Setting the `base_url` and `api_key` directly in `CartoContext`. This
+         method is easier.::
+
+            cc = CartoContext(
+                base_url='https://eschbacher.carto.com',
+                api_key='abcdefg')
+
+      2. By passing a :obj:`Credentials` instance in `CartoContext`'s `creds`
+         keyword argument. This method is more flexible.::
+
+            from cartoframes import Credentials
+            creds = Credentials(user='eschbacher', key='abcdefg')
+            cc = CartoContext(creds=creds)
 
     Attributes:
-        creds (cartoframes.Credentials): :obj:`Credentials` instance
+        creds (:obj:`Credentials`): :obj:`Credentials` instance
 
     Args:
         base_url (str): Base URL of CARTO user account. Cloud-based accounts
@@ -82,6 +98,8 @@ class CartoContext(object):
             a personal or multi-user account. On-premises installation users
             should ask their admin.
         api_key (str): CARTO API key.
+        creds (:obj:`Credentials`): A :obj:`Credentials` instance can be used
+          in place of a `base_url`/`api_key` combination.
         session (requests.Session, optional): requests session. See `requests
             documentation
             <http://docs.python-requests.org/en/master/user/advanced/>`__
@@ -176,7 +194,7 @@ class CartoContext(object):
               lnglat=None, encode_geom=False, geom_col=None, **kwargs):
         """Write a DataFrame to a CARTO table.
 
-        Example:
+        Examples:
             Write a pandas DataFrame to CARTO.
 
             .. code:: python
@@ -219,7 +237,8 @@ class CartoContext(object):
                 as `the_geom`.
             geom_col (str, optional): The name of the column where geometry
                 information is stored. Used in conjunction with `encode_geom`.
-            kwargs: Keyword arguments to control write operations. Options are:
+            **kwargs: Keyword arguments to control write operations. Options
+                are:
 
                 - `compression` to set compression for files sent to CARTO.
                   This will cause write speedups depending on the dataset.
@@ -669,19 +688,18 @@ class CartoContext(object):
             #  bug is fixed ref: support/1127
             try:
                 self.sql_client.send('''
-                    create table {0} as SELECT 1;
-                    drop table {0};
+                    CREATE TABLE {0} AS SELECT 1;
+                    DROP TABLE {0};
                 '''.format(table_name))
                 resp = self._auth_send(
                     'api/v1/imports', 'POST',
-                    params=dict(sql=query,
-                                # collision_strategy='',
-                                table_name=table_name),
+                    params=dict(table_name=table_name),
+                    json=dict(sql=query),
+                    # collision_strategy='',
                     headers={'Content-Type': 'application/json'})
-            except CartoException:
+            except CartoException as err:
                 raise CartoException(
-                    'Table `{0}` already exists. Delete it before creating a '
-                    'table from this query'.format(table_name))
+                    'Cannot create table `{0}`: {1}'.format(table_name, err))
 
             while True:
                 import_job = self._check_import(resp['item_queue_id'])
@@ -1625,17 +1643,20 @@ class CartoContext(object):
             **DEFAULT_SQL_ARGS
         )['fields'].keys()
 
-        if set(tablecols) & set(_meta['suggested_name']):
-            commoncols = set(tablecols) & set(_meta['suggested_name'])
-            raise NameError('Column name collision for column(s): {cols}. '
-                            'Rename table column(s) to resolve.'.format(
-                                cols=', '.join(commoncols)))
+        names = {}
+        for suggested in _meta['suggested_name']:
+            if suggested in tablecols:
+                names[suggested] = utils.unique_colname(suggested, tablecols)
+                warn('{s0} was augmented as {s1} because of name collision'. \
+                    format(s0=suggested, s1=names[suggested]))
+            else:
+                names[suggested] = suggested
 
         cols = ', '.join(
             '(data->{n}->>\'value\')::{pgtype} AS {col}'.format(
                 n=row[0],
                 pgtype=row[1]['numer_type'],
-                col=row[1]['suggested_name'])
+                col=names[row[1]['suggested_name']])
             for row in _meta.iterrows())
         query = utils.minify_sql((
             'SELECT t.*, {cols}',
@@ -1668,7 +1689,10 @@ class CartoContext(object):
         res = self.auth_client.send(relative_path, http_method, **kwargs)
         if isinstance(res.content, str):
             return json.loads(res.content)
-        return json.loads(res.content.decode('utf-8'))
+        try:
+            return json.loads(res.content.decode('utf-8'))
+        except json.JSONDecodeError as err:
+            raise CartoException(err)
 
     def _check_query(self, query, style_cols=None):
         """Checks if query from Layer or QueryLayer is valid"""